Actually added four player views to the game

+++ core/src/com/mygdx/game/subsystems/Subsystem.java
@@ -2,5 +2,5 @@ package com.mygdx.game.subsystems;
 
 public interface Subsystem
 {
-    void update(float deltaInMillis);
+    void update(long deltaInMillis);
 }
